requirements.txt
google-cloud-vision
pillow